slighty reduced memory footprint on CPU side and streamlined int types

+++ FontLibrary/Font.h
@@ -20,23 +20,23 @@ enum SDFTYPE
 class CCharInfo
 {
 public:
-	unsigned int m_ID = 0;
-	unsigned int m_textureX = 0;
-	unsigned int m_textureY = 0;
-	unsigned int m_textureHeight = 0;
-	unsigned int m_textureWidth = 0;
-	int m_xoffset = 0;
-	int m_yoffset = 0;
-	int m_xadvance = 0;
-	//unsigned int m_page = 0; //unused, kept for future development
-	//unsigned int m_channel = 0; //unused, kept for future development
+	uint32_t m_ID = 0;
+	uint16_t m_textureX = 0;
+	uint16_t m_textureY = 0;
+	uint16_t m_textureHeight = 0;
+	uint16_t m_textureWidth = 0;
+	int16_t m_xoffset = 0;
+	int16_t m_yoffset = 0;
+	uint16_t m_xadvance = 0;
+	//uint16_t m_page = 0; //unused, kept for future development
+	//uint16_t m_channel = 0; //unused, kept for future development
 };
 
 class CKerning
 {
 public:
-	unsigned int m_first = 0;
-	unsigned int m_second = 0;
+	uint16_t m_first = 0;
+	uint16_t m_second = 0;
 	float m_amount = 0.0;
 };
 
@@ -49,20 +49,20 @@ public:
 
 	//variables:
 	std::string m_face;
-	unsigned int m_size = 0;
+	uint16_t m_size = 0;
 	bool	m_bold = false;
 	bool	m_italic = false;
 	bool	m_unicode = false;
 	std::string m_bitmapFileName;
 	CCharInfo** m_fontCharInfo = nullptr;
-	unsigned int m_highestASCIIChar = 0;
-	unsigned int m_textureH = 0;
-	unsigned int m_textureW = 0;
+	uint32_t m_highestASCIIChar = 0;
+	uint16_t m_textureH = 0;
+	uint16_t m_textureW = 0;
 	float**	m_kernings = nullptr;
-	unsigned int m_highestKerningFirst = 0;
-	unsigned int m_highestKerningSecond = 0;
-	unsigned int m_fontTextures = 0;
-	int		m_base = 0;
+	uint32_t m_highestKerningFirst = 0;
+	uint32_t m_highestKerningSecond = 0;
+	uint32_t m_fontTextures = 0;
+	uint16_t	m_base = 0;
 	SDFTYPE m_sdfType = SDF_NO;
-	int		m_sdfRange = 4; //only used for multichannel SDFs
+	uint8_t	m_sdfRange = 4; //only used for multichannel SDFs
 };
